backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cffi==1.15.0
coverage==6.1.1
cryptography==35.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.8.0
flake8==4.0.1
importlib-metadata==4.2.0
mccabe==0.6.1
Pillow==8.4.0
pkg_resources==0.0.0
psycopg2==2.9.1
pycodestyle==2.8.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
uWSGI==2.0.19.1
zipp==3.6.0
